[FIXED] Axe drifting bug & Tree Canvas Scale Bug
There were two bugs that I fixed :elephant: The axe drifting was caused by some script changing the position of the axe. The fix I did isn't the greatest, but I'm just overwriting the position and rotation at the end of the frame. The Tree script wasn't taking into account the parent's scale. So I just had to do a quick division to get that consistent with all the trees